requirements.txt
python-decouple==3.6
selenium==4.5.0
